+++ CHANGELOG.md
@@ -36,6 +36,8 @@ Currently, this repo is in Prerelease. When it is released, this project will ad
 - Adds snapshot tests for the `Accordion` component, as well as `id` props to the components within the `Accordion` so snapshot tests pass.
 - Passes an `id` to the `Icon` in the `Select` component.
 - Adds snapshot tests for the `Select` componnet.
+- Updates font size to "12px" and top margin to "4px" for `HelperErrorText` component.
+- Updates font size to "14px" for `TextInput` component.
 
 ### Fixes